+++ eslint.config.js
@@ -1,12 +1,12 @@
 import js from "@eslint/js";
-import globals from "globals";
 import reactHooks from "eslint-plugin-react-hooks";
 import reactRefresh from "eslint-plugin-react-refresh";
-import tseslint from "typescript-eslint";
 import { globalIgnores } from "eslint/config";
+import globals from "globals";
+import tseslint from "typescript-eslint";
 
 export default tseslint.config([
-  globalIgnores(["dist"]),
+  globalIgnores(["dist", "src/components/ui/**"]),
   {
     files: ["**/*.{ts,tsx}"],
     extends: [
@@ -15,6 +15,7 @@ export default tseslint.config([
       reactHooks.configs["recommended-latest"],
       reactRefresh.configs.vite,
     ],
+    ignores: ["src/components/ui/**"],
     languageOptions: {
       ecmaVersion: 2020,
       globals: globals.browser,
